+++ sdl3-ttf-sys/ChangeLog.md
@@ -6,6 +6,9 @@
     - Add metadata
     - Remove deprecated aliases
 
+- next:
+    - Expose `TTF_TextData`
+
 - 0.2.0:
     - Update sdl3-sys to 0.5.0
     - Add `no-link` feature